requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
astkit==0.5.4
canonical==0.52.0
certifi==2024.7.4
cffi==1.17.0
ckms==0.65.0
click==8.1.7
colorama==0.4.6
crcmod==1.7
cryptography==43.0.0
dnspython==2.6.1
dsnparse==0.2.1
email_validator==2.2.0
exceptiongroup==1.2.2
flexcache==0.3
flexparser==0.3.1
future==1.0.0
h11==0.14.0
headless==1.12.4
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflect==7.3.1
instrumental==0.5.3
Instrumental-lib==0.7
iso3166==2.1.1
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
more-itertools==10.4.0
NiceLib==0.7.1
numpy==2.0.1
opencv-python==4.10.0.84
phonenumbers==8.13.42
Pint==0.24.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyrealsense2==2.55.1.6486
pytz==2024.1
pywin32==306
PyYAML==6.0.2
scipy==1.14.0
sniffio==1.3.1
termcolor==2.4.0
typeguard==4.3.0
typing_extensions==4.12.2
unimatrix==1.14.0
xArm-Python-SDK==1.13.8